requirements.txt
asyncpg==0.27.0
discord.py==2.1.0
pytz==2022.7.1
sqlalchemy[asyncio]